Pass props up
I had been looking around googling for the method of passing props back up to higher level componeneta and I finally found it in this video!! https://youtu.be/A71aqufiNtQ 

It's about 34 min in. Now I can move the nav bar and pass up a Boolean to activate a state that changes the margin style of the app!