img added to Home.js
img added to home to play with using it as a background image. Works well so i'm going to use that same method for about and contact. Currently trying to locate some SVG files to use as a background so I can add some interesting effects or animations if I am able.